requirements.txt
fastapi[standard]
pydantic
tortoise-orm
asyncpg
aerich
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-multipart